fix: prevent integer overflow in token counts and add validation
- Change token counts from int to qint64 in CostUsageReport.h structs
  to handle users with billions of tokens
- Add PRICING_VERSION to cache to invalidate when pricing constants change
- Bump CACHE_SCHEMA_VERSION from 1 to 2 for qint64 serialization
- Add validateTokens() and validateCost() to catch corrupted data
  (negative values, unreasonably large values)
- Apply validation at all JSON parsing locations in scanClaude(),
  scanCodex(), scanPi(), and scanOpenCodeDB()
- Improve model name matching with word-boundary-like checks to avoid
  false matches (e.g., "agpt-xxx" no longer matches "gpt")

+++ src/models/CostUsageReport.h
@@ -3,33 +3,34 @@
 #include <QString>
 #include <QVector>
 #include <QDateTime>
+#include <QtGlobal>
 #include <optional>
 
 struct CostUsageModelBreakdown {
     QString modelName;
-    int inputTokens = 0;
-    int cacheReadTokens = 0;
-    int cacheCreationTokens = 0;
-    int outputTokens = 0;
-    int totalTokens() const { return inputTokens + cacheReadTokens + cacheCreationTokens + outputTokens; }
+    qint64 inputTokens = 0;
+    qint64 cacheReadTokens = 0;
+    qint64 cacheCreationTokens = 0;
+    qint64 outputTokens = 0;
+    qint64 totalTokens() const { return inputTokens + cacheReadTokens + cacheCreationTokens + outputTokens; }
     double costUSD = 0.0;
 };
 
 struct CostUsageDailyEntry {
     QString date;
-    int inputTokens = 0;
-    int cacheReadTokens = 0;
-    int cacheCreationTokens = 0;
-    int outputTokens = 0;
-    int totalTokens() const { return inputTokens + cacheReadTokens + cacheCreationTokens + outputTokens; }
+    qint64 inputTokens = 0;
+    qint64 cacheReadTokens = 0;
+    qint64 cacheCreationTokens = 0;
+    qint64 outputTokens = 0;
+    qint64 totalTokens() const { return inputTokens + cacheReadTokens + cacheCreationTokens + outputTokens; }
     double costUSD = 0.0;
     QVector<CostUsageModelBreakdown> models;
 };
 
 struct CostUsageSnapshot {
-    int sessionTokens = 0;
+    qint64 sessionTokens = 0;
     double sessionCostUSD = 0.0;
-    int last30DaysTokens = 0;
+    qint64 last30DaysTokens = 0;
     double last30DaysCostUSD = 0.0;
     QVector<CostUsageDailyEntry> daily;
     QDateTime updatedAt;
